Edits

+++ 0x03-python-data_structures/13-is_palindrome.c
@@ -1,6 +1,12 @@
 #include "lists.h"
 
 
+/**
+ * count_listint - returns number of nodes
+ * @h: head of the list
+ * Return: number of nodes
+ */
+
 
 size_t count_listint(const listint_t *h)
 {
@@ -24,7 +30,7 @@ size_t count_listint(const listint_t *h)
 
 /**
  * is_palindrome - checks if a linked list is a palindrome
- * @head: pointer to the head node
+ * @head: pointer to the head pointer
  * Return: 1 if palindrome and 0 if not
  */
 
@@ -34,7 +40,7 @@ int is_palindrome(listint_t **head)
 {
 unsigned int num = count_listint((const listint_t *) *head);
 int num_arr[num / 2];
-size_t i;
+unsigned int i;
 for (i = 0; i < num / 2; i++)
 {
 	num_arr[i] = (*head)->n;
